Use inventory endpoint, explode hostgroup query
Rather than user `/pdb/query/v4/facts` use an inventory query to
`pdb/query/v4`.

Also a change of behaviour previously the hostgroup `a/b/c` was searched
as a regex for `a/b/c` in the hostgroup.

Instead a hostgroup such as `a/b/c` is filtered as the and of
`hostgroup_0=a`, `hostgroup_1=b` and `hostgroup_2=c`.